refactor(api): custom-links → src/api modules (Effort 1)
Mechanical move of the custom-links feature from src/hooks/queries/custom-links to
src/api/custom-links, adding queryOptions factories for the query hooks. No behavior change.
