+++ src/api/custom-links/types.ts
@@ -1,0 +1,9 @@
+import type { Tables } from "@/integrations/supabase/types";
+
+export type CustomLink = Tables<"custom_links">;
+
+export const customLinksKeys = {
+  all: ["customLinks"] as const,
+  byFestival: (festivalId: string) =>
+    [...customLinksKeys.all, festivalId] as const,
+};
