+++ src/api/custom-links/useCustomLinksMutation.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useToast } from "@/hooks/use-toast";
 import { supabase } from "@/integrations/supabase/client";
-import { customLinksKeys } from "./useCustomLinks";
+import { customLinksKeys } from "./types";
 
 interface BulkUpdateCustomLinksData {
   festivalId: string;
